backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase>=2.8.0
python-dotenv==1.0.0
httpx>=0.27.0,<0.28.0
websockets>=13.0,<16
